Minor fixes and updates.

+++ extern/sdl4ogre/sdlcursormanager.cpp
@@ -110,7 +110,7 @@ namespace SFO
                 Ogre::ColourValue clr = destImage.getColourAt(x, y, 0);
 
                 //set the pixel on the SDL surface to the same value as the Ogre texture's
-                _putPixel(surf, x, y, SDL_MapRGBA(surf->format, static_cast<Uint8>(clr.r * 255), 
+                _putPixel(surf, (int)x, (int)y, SDL_MapRGBA(surf->format, static_cast<Uint8>(clr.r * 255), 
                     static_cast<Uint8>(clr.g * 255), static_cast<Uint8>(clr.b * 255), static_cast<Uint8>(clr.a * 255)));
             }
         }
